Refactored payer-patient elgibility criteria into a stategy algorithm format.
