+++ src/main/java/org/mitre/synthea/world/agents/PayerController.java
@@ -14,13 +14,13 @@ import java.util.stream.Collectors;
 import org.mitre.synthea.helpers.Config;
 import org.mitre.synthea.helpers.SimpleCSV;
 import org.mitre.synthea.helpers.Utilities;
-import org.mitre.synthea.world.agents.behaviors.IPayerAdjustment;
-import org.mitre.synthea.world.agents.behaviors.IPayerFinder;
-import org.mitre.synthea.world.agents.behaviors.PayerAdjustmentFixed;
-import org.mitre.synthea.world.agents.behaviors.PayerAdjustmentNone;
-import org.mitre.synthea.world.agents.behaviors.PayerAdjustmentRandom;
-import org.mitre.synthea.world.agents.behaviors.PayerFinderBestRates;
-import org.mitre.synthea.world.agents.behaviors.PayerFinderRandom;
+import org.mitre.synthea.world.agents.behaviors.payer_adjustment.IPayerAdjustment;
+import org.mitre.synthea.world.agents.behaviors.payer_adjustment.PayerAdjustmentFixed;
+import org.mitre.synthea.world.agents.behaviors.payer_adjustment.PayerAdjustmentNone;
+import org.mitre.synthea.world.agents.behaviors.payer_adjustment.PayerAdjustmentRandom;
+import org.mitre.synthea.world.agents.behaviors.payer_finder.IPayerFinder;
+import org.mitre.synthea.world.agents.behaviors.payer_finder.PayerFinderBestRates;
+import org.mitre.synthea.world.agents.behaviors.payer_finder.PayerFinderRandom;
 import org.mitre.synthea.world.concepts.HealthRecord.EncounterType;
 import org.mitre.synthea.world.geography.Location;
