+++ src/main/java/org/mitre/synthea/world/agents/Provider.java
@@ -24,10 +24,10 @@ import org.mitre.synthea.helpers.Config;
 import org.mitre.synthea.helpers.RandomNumberGenerator;
 import org.mitre.synthea.helpers.SimpleCSV;
 import org.mitre.synthea.helpers.Utilities;
-import org.mitre.synthea.world.agents.behaviors.IProviderFinder;
-import org.mitre.synthea.world.agents.behaviors.ProviderFinderNearest;
-import org.mitre.synthea.world.agents.behaviors.ProviderFinderQuality;
-import org.mitre.synthea.world.agents.behaviors.ProviderFinderRandom;
+import org.mitre.synthea.world.agents.behaviors.provider_finder.IProviderFinder;
+import org.mitre.synthea.world.agents.behaviors.provider_finder.ProviderFinderNearest;
+import org.mitre.synthea.world.agents.behaviors.provider_finder.ProviderFinderQuality;
+import org.mitre.synthea.world.agents.behaviors.provider_finder.ProviderFinderRandom;
 import org.mitre.synthea.world.concepts.ClinicianSpecialty;
 import org.mitre.synthea.world.concepts.HealthRecord.EncounterType;
 import org.mitre.synthea.world.concepts.Names;
