+++ src/main/java/org/mitre/synthea/world/agents/behaviors/provider_finder/ProviderFinderQuality.java
@@ -1,4 +1,4 @@
-package org.mitre.synthea.world.agents.behaviors;
+package org.mitre.synthea.world.agents.behaviors.provider_finder;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,16 +7,25 @@ import org.mitre.synthea.world.agents.Person;
 import org.mitre.synthea.world.agents.Provider;
 import org.mitre.synthea.world.concepts.HealthRecord.EncounterType;
 
-public class ProviderFinderRandom implements IProviderFinder {
+public class ProviderFinderQuality implements IProviderFinder {
+
+  private IProviderFinder nearest = new ProviderFinderNearest();
 
   @Override
   public Provider find(List<Provider> providers, Person person, EncounterType service, long time) {
-    List<Provider> options = new ArrayList<Provider>();
+    List<Provider> options = new ArrayList<>();
+    double bestQuality = Double.NEGATIVE_INFINITY;
 
     for (Provider provider : providers) {
       if (provider.accepts(person, time)
           && (provider.hasService(service) || service == null)) {
-        options.add(provider);
+        if (provider.quality > bestQuality) {
+          options.clear();
+          options.add(provider);
+          bestQuality = provider.quality;
+        } else if (provider.quality == bestQuality) {
+          options.add(provider);
+        }
       }
     }
 
@@ -25,8 +34,7 @@ public class ProviderFinderRandom implements IProviderFinder {
     } else if (options.size() == 1) {
       return options.get(0);
     } else {
-      // there are a few equally good options, pick one randomly.
-      return options.get(person.randInt(options.size()));
+      return nearest.find(options, person, service, time);
     }
   }
 }
