+++ src/test/java/org/mitre/synthea/world/agents/behaviors/PayerFinderTest.java
@@ -14,6 +14,8 @@ import org.mitre.synthea.helpers.Config;
 import org.mitre.synthea.world.agents.Payer;
 import org.mitre.synthea.world.agents.PayerController;
 import org.mitre.synthea.world.agents.Person;
+import org.mitre.synthea.world.agents.behaviors.payer_finder.PayerFinderBestRates;
+import org.mitre.synthea.world.agents.behaviors.payer_finder.PayerFinderRandom;
 import org.mitre.synthea.world.geography.Location;
 
 public class PayerFinderTest {
